pyproject.toml
[project]
name = "bridgemaster-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "endplay>=0.5.12",
    "google-genai>=1.56.0",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
]
